Оставшиеся фиксы поиска по айди + фиксы тестов и связанного кода
