+++ src/test/java/faang/school/postservice/service/CommentServiceTest.java
@@ -74,7 +74,7 @@ public class CommentServiceTest {
                 .id(postId)
                 .build();
 
-        when(postService.findPostById(postId)).thenReturn(post);
+        when(postService.findPostById(postId)).thenReturn(Optional.ofNullable(post));
         when(commentMapper.toEntity(commentDto)).thenReturn(comment);
 
         // act
